Revert "Added the card action manager to the scene as a singleton but I think each player might need their own unsure"
This reverts commit a8404989a3cfdbe5d5a4ca5f0ad9893f500fa7a7.

+++ ResilienceGame/Assets/Cards/CardData.cs
@@ -59,9 +59,9 @@ public struct CardData
     public float percentSuccess;
     public int cardID;
     public int teamID; // TODO: Use for SD
-    public int blueCost;
-    public int blackCost;
-    public int purpleCost;
+    public float blueCost;
+    public float blackCost;
+    public float purpleCost;
     public int drawAmount;
     public int removeAmount;
     public int targetAmount;
@@ -72,6 +72,6 @@ public struct CardData
     public int duration;
     public bool hasDoomEffect;
     public string[] meepleType;
-    public int meepleAmount;
+    public float meepleAmount;
     public PlayerSector[] onlyPlayedOn;
 }
